// src/app/core/config/data.state.enum.ts
export enum DataStateEnum {
  LOADING="Loading",
  SUCCESS="success",
  ERROR="Error",
  INITIAL="Initial",
}

export enum OperationEnum {
  CREATE = "CREATE",
  UPDATE = "UPDATE",
  ADD = "ADD",
  ERROR = "ERROR",
  LOOSING = "LOOSING",
  INTIATED = "INTIATED",
  INITIATED = "INITIATED",
  initiated = "Initialized",
  WINNING = "WINNING",
  PENDING = "PENDING",
  SUCCESS = "SUCCESS"
  //  FAILED  Failed
}

export enum StatusEnum {
  ACTIVATE = "ACTIVE",
  DESACTIVATE = "INACTIVE",
  ALL = 'ALL',

}

export enum AddressEnum {
  TYPE = "LOCATION",
  UNKNOWN = "UNKNOWN"
}

export enum StateAddressEnum {
  ACTIVE = "ACTIVE",
  CREATE_BUT_NOT_ACTIVE = "CREATE_BUT_NOT_ACTIVE",
  DELETED = "DELETED"
}

export enum RoleEnum {
  ACTIVE = "ACTIVE",
  CREATE_BUT_NOT_ACTIVE = "CREATE_BUT_NOT_ACTIVE",
  DELETED = "DELETED"
}

export enum UserStateEnum { 
  ALL = 'ALL',
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  TREATMENT = 'TREATMENT',
  BLOCKED = 'BLOCKED',
  CREATE_BUT_NOT_ACTIVE = 'CREATE_BUT_NOT_ACTIVE',
  DELETED = 'DELETED',
}

export enum GroupCodeEnum {
  MANAGE_ADDRESS = "MANAGE_ADDRESS",
  MANAGE_AUTHORISATIONS = "MANAGE_AUTHORISATIONS",
  MANAGE_ROLE = "MANAGE_ROLE",
  MANAGE_SAAS_USER = "MANAGE_SAAS_USER",
  MANAGE_USER = "MANAGE_USER",
  NOT_SPECIFY = "NOT_SPECIFY"
}


export enum AuthorisationKeyEnum {
  ADD_DOCUMENT_TO_USER = "ADD_DOCUMENT_TO_USER",
  BLOCK_USER = "BLOCK_USER",
  CREATE_ADDRESS = "CREATE_ADDRESS",
  CREATE_ROLE = "CREATE_ROLE",
  CREATE_SAAS_USER = "CREATE_SAAS_USER",
  CREATE_USER = "CREATE_USER",
  DELETE_ADDRESS = "DELETE_ADDRESS",
  DELETE_ROLE = "DELETE_ROLE",
  DELETE_SAAS_USER = "DELETE_SAAS_USER",
  DELETE_USER = "DELETE_USER",
  GET_ALL_ADDRESS = "GET_ALL_ADDRESS",
  GET_ALL_AUTHORISATIONS = "GET_ALL_AUTHORISATIONS",
  GET_ALL_SAAS_USERS = "GET_ALL_SAAS_USERS",
  GET_DOCUMENT_OF_USER = "GET_DOCUMENT_OF_USER",
  GET_MY_ADDRESS = "GET_MY_ADDRESS",
  GET_ROLE = "GET_ROLE",
  GET_SAAS_USER = "GET_SAAS_USER",
  GET_USERS = "GET_USERS",
  GET_USER_DETAILS = "GET_USER_DETAILS",
  NOT_SPECIFY = "NOT_SPECIFY",
  REMOVE_DOCUMENT_TO_USER = "REMOVE_DOCUMENT_TO_USER",
  UNBLOCK_USER = "UNBLOCK_USER",
  UPDATE_ADDRESS = "UPDATE_ADDRESS",
  UPDATE_MY_INFOS = "UPDATE_MY_INFOS",
  UPDATE_PASSWORD = "UPDATE_PASSWORD",
  UPDATE_ROLE = "UPDATE_ROLE",
  UPDATE_SAAS_USER = "UPDATE_SAAS_USER",
  UPDATE_USER = "UPDATE_USER"
}

export enum SortEnum {
  ASC = "asc",
  DESC = "desc"
}

export enum TypeGrilleEnum {
  ECONOMIQUE = "ECONOMIQUE",
  STANDARD = "STANDARD",
  CONFORT = "CONFORT",
  PRESTIGE = "PRESTIGE",
  PERSONALISEE = "PERSONALISEE"
}

export enum AvisStateEnum {
  APPROUVE = "APPROUVE",
  EN_COURS = "EN_COURS",
  REJETE = "REJETE"
}